feat(asterisk): support srtp
libsrtp was missing which meant res_srtp.so was not available after
compilation. It has been added and the moduel is now loaded by default
which makes it possible to encrypt calls.

Tasks are also now separated by tags which lets users either run the
asterisk or crowdsec section independently when running the
asterisk-servers playbook.